notes/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.3.1
MarkupSafe==2.1.1
oauthlib==3.2.0
pycparser==2.21
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
